customer_support_agent/pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.12",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.2",
]
